Add NYT article re Israel and Russia Kaspersky backdoor
NYT article about Kaspersky indicates that AV are generic backdoors into systems, given their requirement for full privilege. Not modifying the "will likely increase attack surface" text, but probably useful to have a discussion about that part also with a view to changing it to a direct recommendation to avoid.